refactor: 데이터 관리 로직 분리
